#272 implement default spatial methods
